Refactor out `lovatAPI.getScoutReportAnalysis()`
